requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2